pyre setup
