+++ requirements.txt
@@ -5,10 +5,12 @@ hypothesis
 iopath
 numpy
 pandas
-pyre-extensions==0.0.27
+pyre-check-nightly==0.0.101675080701
+pyre-extensions
 scikit-build
 tabulate
 torcharrow
 torchmetrics
 torchx
 tqdm
+usort
